+++ assets/assets.go
@@ -1,0 +1,6 @@
+package assets
+
+import _ "embed"
+
+//go:embed ascii-art.txt
+var ASCIILogo string